OPTM-113 Moved task list service related files to separate folder
