+++ lib/screens/task_list_form_screen.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:p2p_task/models/task_list.dart';
 import 'package:p2p_task/services/change_callback_notifier.dart';
-import 'package:p2p_task/services/task_list_service.dart';
+import 'package:p2p_task/services/task_list/task_list_service.dart';
 import 'package:provider/provider.dart';
 
 class TaskListFormScreen extends StatefulWidget {
